backend/requirements.txt
# Core API
fastapi>=0.130.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Data science (no pinned versions — use latest with Python 3.14 wheels)
networkx>=3.4
scikit-learn>=1.5
pandas>=2.2
numpy>=2.1
joblib>=1.4

# ML — XGBoost + SHAP explainability
xgboost>=2.1
shap>=0.46

# Google AI (free tier Gemini API)
google-generativeai>=0.8.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23

